move github stuff under the github.py module
